+++ actioncable/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Respect calls to `#reject` in `#subscribe` and its callbacks.
+
+    *Joshua Young*
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Add an `identifier` to the event payload for the ActiveSupport::Notification `transmit_subscription_confirmation.action_cable` and `transmit_subscription_rejection.action_cable`.